// nodejs/semantic-kernel/connectors/ai/open-ai/services/open-ai-model-types.ts
/**
 * OpenAI model types, can be text, chat, embedding, etc.
 */
export enum OpenAIModelTypes {
  TEXT = 'text',
  CHAT = 'chat',
  EMBEDDING = 'embedding',
  TEXT_TO_IMAGE = 'text-to-image',
  AUDIO_TO_TEXT = 'audio-to-text',
  TEXT_TO_AUDIO = 'text-to-audio',
  REALTIME = 'realtime',
  RESPONSE = 'response',
}